Update preview-frame loading logic to support chrome.
It appears that chrome has slightly different timing for loading
iframe srcdocs. It appears that the iframe's `contentDocument` was
not updated by the time that `copyFromHTML` was being called, when
switching back to a tab.

I've updated the code to use always wait on the *iframe*'s load event,
which we listen on *before* setting the srcdoc attribute.