Remove the use of eslint-disable-next-line for fallthrough.
The eslint rule `no-fallthrough` looks for the word `fallthrough`on the
previous line to disable the warning when deliberately falling
through. Thus using
```
// eslint-disable-next-line no-fallthrough
```
causes the lint to not trigger (beacause it has the word fallthrough), as well
as disables that lint. This causes the `eslint-disable-next-line` to be
considered unused.

This also enables `reportUnusedDisableDirectives`, which would have reported
this error.